Thang

+++ CoreBase/Test/Dictionary/frmService.cs
@@ -226,7 +226,17 @@ namespace AusNail.Dictionary
             if (result == DialogResult.Yes)
             {
                 this.zDeleteRow = ((DataTable)Bds.DataSource).Rows[_postion];
-                bool flag = base.DeleteData();
+                // Check use for bill or booking.
+                DataTable dataTable = MsSqlHelper.ExecuteDataTable(ZenDatabase.ConnectionString, "zCheckServiceExists", int.Parse(zDeleteRow["ServiceID"].ToString()));
+                if (dataTable != null && dataTable.Rows.Count > 0)
+                {
+                    MessageBox.Show("Service is already in use, cannot be deleted.!", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    return;
+                }
+                else
+                {
+                    bool flag = base.DeleteData();
+                }
                 LoadData();
             }
             else
